Remove buffer size restrictions (malloc_ver)
Some students told me BUFFER_SIZE must not be restricted to avoid stack overflows, instead, malloc should be used. I think this is bad for the program as malloc can fail at some point for any reason, and the program that grades this project doesn't check if this is done and may even fail you if malloc fails as it expects GNL to return a line and not NULL, plus, you shouldn't be using a BUFFER_SIZE high enough to overflow the stack anyways. That's the reason why this "fix" is included on a separate folder, malloc_ver, just for those who want it.